Switch to newer epics_device release with support for waveform length
We're going to be using the newer version of epics_device, but there's an
incompatible change the waveform interface which needs to be fixed first.

+++ epics/src/adc.c
@@ -30,7 +30,7 @@ static struct adc_context {
 } adc_context[AXIS_COUNT];
 
 
-static void write_adc_taps(void *context, float array[], size_t *length)
+static void write_adc_taps(void *context, float array[], unsigned int *length)
 {
     struct adc_context *adc = context;
     *length = hardware_config.adc_taps;
